+++ src/index.js
@@ -1,7 +1,11 @@
 export { default as TableCore } from './components/TableCore';
+export { default as TablePagination } from "./components/TablePagination";
 export { default as TableActions } from "./models/actions";
 export { default as createTable } from "./store/table";
 export { default as eventMiddleware } from './store/eventMiddleware';
+export { default as useTableStoreHooks } from "./hooks/useTableStoreHooks";
+export { default as usePagination } from "./hooks/usePagination";
 export { makeGetPageCount } from "./selectors/paginationSelectors";
 export { setDefaultOptions } from './utils/optionUtils';
-export { getTableSlice } from "./utils/reduxUtils";
+export { getTableSlice } from "./utils/optionUtils";
+